git-svn-id: https://repo.thevirtualbrain.org/svn/tvb/trunk@9292 1c0e02f0-7929-43c0-8fb3-3293bf43b0d1

+++ framework_tvb/tvb/interfaces/web/controllers/settings_controller.py
@@ -174,7 +174,7 @@ class DiskSpaceValidator(formencode.FancyValidator):
         :param value is user-specified value, in MB
         """
         try:
-            value = long(value)
+            value = int(value)
             return value
         except ValueError:
             raise formencode.Invalid('Invalid disk space %s. Should be number' % value, value, None)
